requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
chromadb
sentence-transformers
